第12次提交，实现购物车下单

+++ shop_service_api/shop_service_goods_api/src/main/java/cn/lx/shop/goods/pojo/Sku.java
@@ -33,7 +33,7 @@ public class Sku implements Serializable{
 
 	@ApiModelProperty(value = "价格（分）",required = false)
     @Column(name = "price")
-	private Integer price;//价格（分）
+	private Double price;//价格（分）
 
 	@ApiModelProperty(value = "库存数量",required = false)
     @Column(name = "num")
@@ -129,12 +129,12 @@ public class Sku implements Serializable{
 		this.name = name;
 	}
 	//get方法
-	public Integer getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
 	//set方法
-	public void setPrice(Integer price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 	//get方法
